+++ src/org/sosy_lab/java_smt/delegate/statistics/SolverStatistics.java
@@ -36,6 +36,7 @@ public class SolverStatistics {
   final AtomicInteger bvOperations = new AtomicInteger();
   final AtomicInteger fpOperations = new AtomicInteger();
   final AtomicInteger typeOperations = new AtomicInteger();
+  final AtomicInteger stringOperations = new AtomicInteger();
 
   // model operations
   final AtomicInteger modelEvaluations = new AtomicInteger();
@@ -138,6 +139,10 @@ public class SolverStatistics {
 
   public int getNumberOfFPOperations() {
     return fpOperations.get();
+  }
+
+  public int getNumberOfStringOperations() {
+    return stringOperations.get();
   }
 
   public int getNumberOfModelEvaluationQueries() {