+++ src/org/sosy_lab/java_smt/solvers/z3/Z3SolverContext.java
@@ -154,6 +154,10 @@ public final class Z3SolverContext extends AbstractSolverContext {
     Native.incRef(context, Native.sortToAst(context, integerSort));
     long realSort = Native.mkRealSort(context);
     Native.incRef(context, Native.sortToAst(context, realSort));
+    long stringSort = Native.mkStringSort(context);
+    Native.incRef(context, Native.sortToAst(context, stringSort));
+    long regexSort = Native.mkReSort(context, stringSort);
+    Native.incRef(context, Native.sortToAst(context, regexSort));
 
     // The string representations of Z3s formulas should be in SMTLib2,
     // otherwise serialization wouldn't work.
@@ -165,7 +169,15 @@ public final class Z3SolverContext extends AbstractSolverContext {
         context, z3params, Native.mkStringSymbol(context, ":random-seed"), (int) randomSeed);
 
     Z3FormulaCreator creator =
-        new Z3FormulaCreator(context, boolSort, integerSort, realSort, config, pShutdownNotifier);
+        new Z3FormulaCreator(
+            context,
+            boolSort,
+            integerSort,
+            realSort,
+            stringSort,
+            regexSort,
+            config,
+            pShutdownNotifier);
 
     // Create managers
     Z3UFManager functionTheory = new Z3UFManager(creator);
@@ -180,6 +192,7 @@ public final class Z3SolverContext extends AbstractSolverContext {
         new Z3FloatingPointFormulaManager(creator, pFloatingPointRoundingMode);
     Z3QuantifiedFormulaManager quantifierManager = new Z3QuantifiedFormulaManager(creator);
     Z3ArrayFormulaManager arrayManager = new Z3ArrayFormulaManager(creator);
+    Z3StringFormulaManager stringTheory = new Z3StringFormulaManager(creator);
 
     // Set the custom error handling
     // which will throw Z3Exception
@@ -196,7 +209,8 @@ public final class Z3SolverContext extends AbstractSolverContext {
             bitvectorTheory,
             floatingPointTheory,
             quantifierManager,
-            arrayManager);
+            arrayManager,
+            stringTheory);
     return new Z3SolverContext(
         creator, config, z3params, pShutdownNotifier, logger, manager, solverLogfile);
   }