+++ src/org/sosy_lab/java_smt/test/ModelTest.java
@@ -654,6 +654,18 @@ public class ModelTest extends SolverBasedTest0 {
   }
 
   @Test
+  public void testGetString() throws SolverException, InterruptedException {
+    requireStrings();
+    for (String word : new String[] {"", "a", "abc", "1", "123", "-abc", "\"test\"", "\""}) {
+      testModelGetters(
+          smgr.equal(smgr.makeVariable("x"), smgr.makeString(word)),
+          smgr.makeVariable("x"),
+          word,
+          "x");
+    }
+  }
+
+  @Test
   public void testGetModelAssignments() throws SolverException, InterruptedException {
     if (imgr != null) {
       testModelIterator(
@@ -2065,11 +2077,6 @@ public class ModelTest extends SolverBasedTest0 {
     requireParser();
     requireArrays();
     requireBitvectors();
-
-    assume()
-        .withMessage("Solver %s sadly fails on this test.", solverToUse())
-        .that(solverToUse())
-        .isNotEqualTo(Solvers.PRINCESS);
 
     BooleanFormula formula =
         context